--- torchvision/prototype/models/detection/mask_rcnn.py
@@ -1,79 +1,0 @@
-from typing import Any, Optional
-
-from torch import nn
-from torchvision.models._api import WeightsEnum, Weights
-from torchvision.models._meta import _COCO_CATEGORIES
-from torchvision.models._utils import handle_legacy_interface, _ovewrite_value_param
-from torchvision.models.detection.mask_rcnn import (
-    _resnet_fpn_extractor,
-    _validate_trainable_layers,
-    MaskRCNN,
-    misc_nn_ops,
-    overwrite_eps,
-)
-from torchvision.models.resnet import ResNet50_Weights, resnet50
-from torchvision.transforms import ObjectDetectionEval, InterpolationMode
-
-
-__all__ = [
-    "MaskRCNN",
-    "MaskRCNN_ResNet50_FPN_Weights",
-    "maskrcnn_resnet50_fpn",
-]
-
-
-class MaskRCNN_ResNet50_FPN_Weights(WeightsEnum):
-    COCO_V1 = Weights(
-        url="https://download.pytorch.org/models/maskrcnn_resnet50_fpn_coco-bf2d0c1e.pth",
-        transforms=ObjectDetectionEval,
-        meta={
-            "task": "image_object_detection",
-            "architecture": "MaskRCNN",
-            "publication_year": 2017,
-            "num_params": 44401393,
-            "categories": _COCO_CATEGORIES,
-            "interpolation": InterpolationMode.BILINEAR,
-            "recipe": "https://github.com/pytorch/vision/tree/main/references/detection#mask-r-cnn",
-            "map": 37.9,
-            "map_mask": 34.6,
-        },
-    )
-    DEFAULT = COCO_V1
-
-
-@handle_legacy_interface(
-    weights=("pretrained", MaskRCNN_ResNet50_FPN_Weights.COCO_V1),
-    weights_backbone=("pretrained_backbone", ResNet50_Weights.IMAGENET1K_V1),
-)
-def maskrcnn_resnet50_fpn(
-    *,
-    weights: Optional[MaskRCNN_ResNet50_FPN_Weights] = None,
-    progress: bool = True,
-    num_classes: Optional[int] = None,
-    weights_backbone: Optional[ResNet50_Weights] = None,
-    trainable_backbone_layers: Optional[int] = None,
-    **kwargs: Any,
-) -> MaskRCNN:
-    weights = MaskRCNN_ResNet50_FPN_Weights.verify(weights)
-    weights_backbone = ResNet50_Weights.verify(weights_backbone)
-
-    if weights is not None:
-        weights_backbone = None
-        num_classes = _ovewrite_value_param(num_classes, len(weights.meta["categories"]))
-    elif num_classes is None:
-        num_classes = 91
-
-    is_trained = weights is not None or weights_backbone is not None
-    trainable_backbone_layers = _validate_trainable_layers(is_trained, trainable_backbone_layers, 5, 3)
-    norm_layer = misc_nn_ops.FrozenBatchNorm2d if is_trained else nn.BatchNorm2d
-
-    backbone = resnet50(weights=weights_backbone, progress=progress, norm_layer=norm_layer)
-    backbone = _resnet_fpn_extractor(backbone, trainable_backbone_layers)
-    model = MaskRCNN(backbone, num_classes=num_classes, **kwargs)
-
-    if weights is not None:
-        model.load_state_dict(weights.get_state_dict(progress=progress))
-        if weights == MaskRCNN_ResNet50_FPN_Weights.COCO_V1:
-            overwrite_eps(model, 0.0)
-
-    return model
